requirements.txt
asgiref==3.7.2
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
python-decouple==3.8
pyzmq==25.1.2
sqlparse==0.4.4
typing_extensions==4.4.0
tzdata==2023.4
